seperate out single context running logic
